fix1 threads project4

+++ src/main/java/edu/project4/renderers/Renderer.java
@@ -18,7 +18,7 @@ public interface Renderer {
         List<AffineTransformation> affineTransformations,
         int samples,
         short iterationsPerSample,
-        double symmetry
+        int symmetry
     );
 
     default FractalImage render(
